requirements.txt
rasa==3.6.21
openai>=1.0.0
python-dotenv==1.0.1
llama-cpp-python>=0.2.24
sqlalchemy>=1.4.0,<3.0.0
aiohttp==3.9.5
requests==2.32.3
colorama==0.4.6
pydantic>=1.10.9,<3.0.0
PyYAML==6.0.2
streamlit==1.38.0
rasa-sdk>=3.6.0
# --- Added for LangChain & RAG ---
langchain>=0.1.0
langchain-community>=0.0.30
chromadb>=0.4.24
tiktoken>=0.5.1
pdfplumber>=0.10.0
unstructured>=0.12.0
python-docx>=1.0.0